LayoutTests/fast/dom/inline-event-attributes-lookup-removed-form.html

<!DOCTYPE html>
<script src="../js/resources/js-test-pre.js"></script>
<form id="test-form">
    <input id="test-input">
</form>
<script>

description('Tests that lookup in inline event handlers keeps references alive');

function forceGC()
{
    if (window.GCController)
        return GCController.collect();

    // Force garbage collection
    for (var ndx = 0; ndx < 99000; ndx++)
        var str = new String("1234");
}

var result;
var pa = '(FAIL input)'
var ss = '(FAIL form)'

var input = document.getElementById('test-input');
input.pa = 'PA';
input.form.ss = 'SS';
input.setAttribute('onclick', 'result = pa + ss');

var f = input.onclick;

input.parentNode.removeChild(input);
input = null;
forceGC();

f();
shouldBeEqualToString('result', 'PASS');

</script>
<script src="../js/resources/js-test-post.js"></script>
