third_party/blink/web_tests/platform/generic/fast/dom/inline-event-attributes-lookup-removed-form-expected.txt

Tests that lookup in inline event handlers keeps references alive

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is "PASS"
PASS successfullyParsed is true

TEST COMPLETE
